test(channels): rewrite Overview specs for connections popover; drop drawer (EVO-2092)
